requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.8
DateTime==5.1
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
fonttools==4.39.2
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycryptodome==3.17
pyparsing==3.0.9
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.1
regex==2023.3.23
requests==2.28.2
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
zope.interface==6.0
